proper volume column editing

+++ Pattern.cpp
@@ -3,20 +3,21 @@
 #include "Pattern.h"
 
 /*
-* XXGG GGFF FFEE EEEV VVVV VVVI IIII INNN NNNN
+* XHGG GGFF FFEE EEEV VVVV VVVI IIII INNN NNNN
 * 
 * First 7 bits are for the note (0-119 for notes, 127 for note off) [N]
 * Next 6 bits are for the instrument (0-63) [I]
 * Next 7 bits are for the volume (0-64, 127 for blank) [V]
-* Next 5 bits are for the effect type (0-31) [E]
+* Next 5 bits are for the effect type (0-31) [E] 
 * Next 4 bits are for the first effect column [F]
 * Next 4 bits are for the second effect column [G]
+* Next bit determines what digit of the volume paremeter you're editing (editor only) [H]
 */
 
 Pattern::Pattern() {
 	row_count = 64;
 	int size = row_count * MAX_CHANNELS;
-	cells = new int[size];
+	cells = new long long[size];
 	for (int i = 0; i < size; i++) {
 		cells[i] = (NOTE_BLANK | (INSTRUMENT_BLANK << 7) | (VOLUME_BLANK << 13)); 
 	}
@@ -26,11 +27,11 @@ void Pattern::freePattern() {
 	delete[] cells;
 }
 
-void Pattern::setCell(int row, int column, int value) {
+void Pattern::setCell(int row, int column, long long value) {
 	cells[row * MAX_CHANNELS + column] = value;
 }
 
-int Pattern::getCell(int row, int column) {
+long long Pattern::getCell(int row, int column) {
 	return cells[row * MAX_CHANNELS + column];
 }
