+++ Pattern.h
@@ -5,12 +5,12 @@
 class Pattern
 {
 private:
-	int* cells;
+	long long* cells;
 public:
 	Pattern();
 	int row_count;
-	void setCell(int row, int column, int value);
-	int getCell(int row, int column);
+	void setCell(int row, int column, long long value);
+	long long getCell(int row, int column);
 	void setCellNote(int row, int column, int note);
 	int getCellNote(int row, int column);
 	void setCellVolume(int row, int column, int volume);
